Move supabaseClient to lib
